chore: 🤖 CR fixes
