+++ src/components/ExportToCSVModal/ExportToCSVModal.tsx
@@ -1,170 +1,25 @@
-import React, { FC, useContext, useEffect, useState } from 'react';
+import React, { FC, useContext } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { useTranslation } from 'react-i18next';
-import { transparentize } from 'polished';
 
-import {
-  Anchor,
-  Button,
-  Modal,
-  ModalFooter,
-  ModalHeader,
-  Portal,
-  Tabs,
-} from '@keen.io/ui-core';
-import { colors } from '@keen.io/colors';
-import { BodyText } from '@keen.io/typography';
-import { parseQuery } from '@keen.io/parser';
-import { DataExport } from '@keen.io/data-export';
+import { Modal, Portal } from '@keen.io/ui-core';
 
 import {
   dataExportActions,
   dataExportSelectors,
 } from '../../modules/dataExport';
 import { AppContext } from '../../contexts';
-import { getQueryResults } from '../../modules/queries';
-import { exportToCsv } from '../../utils';
-import { getVisualization } from '../../modules/app';
-import { Widget } from '../QueryVisualization/types';
-
-import {
-  ModalBody,
-  FooterButtonsWrapper,
-  TabDescription,
-} from './ExportToCSVModal.styles';
-import { Table } from './components/Table';
+import { ExportToCSV } from './components';
 
 const ExportToCSVModal: FC = () => {
   const dispatch = useDispatch();
   const { modalContainer } = useContext(AppContext);
   const isOpen = useSelector(dataExportSelectors.getExportToCSVModalVisibility);
-  const { t } = useTranslation();
   const onClose = () => dispatch(dataExportActions.showCSVExportModal(false));
-
-  const TabOptions = [
-    {
-      label: t('export_CSV.tabs.visualization_data'),
-      id: 'visualizationData',
-    },
-    {
-      label: t('export_CSV.tabs.raw_data'),
-      id: 'rawData',
-    },
-  ];
-
-  const [activeTab, setActiveTab] = React.useState(TabOptions[0].id);
-
-  const [rawData, setRawData] = useState([[]]);
-  const [visualizationData, setVisualizationData] = useState([[]]);
-
-  const queryResults = useSelector(getQueryResults);
-  const { type: widgetType, chartSettings } = useSelector(getVisualization);
-
-  useEffect(() => {
-    if (queryResults) {
-      const { data, keys } = parseQuery(queryResults);
-      setRawData(DataExport.exportRawData({ keys, data }));
-      setVisualizationData(
-        DataExport.exportVisualizationData({
-          query: queryResults.query,
-          chartSettings: {
-            ...chartSettings,
-            data,
-            keys,
-          },
-          widgetType: widgetType as Widget,
-        })
-      );
-    }
-  }, [activeTab, widgetType, queryResults]);
-
-  const onExportToCSV = () => {
-    if (activeTab === 'visualizationData') {
-      exportToCsv({ data: visualizationData });
-    } else if (activeTab === 'rawData') {
-      exportToCsv({ data: rawData });
-    }
-    dispatch(dataExportActions.showCSVExportModal(false));
-  };
 
   return (
     <Portal modalContainer={modalContainer}>
       <Modal isOpen={isOpen} onClose={onClose}>
-        {() => (
-          <>
-            <ModalHeader onClose={onClose}>
-              {t('export_CSV.modal_title')}
-            </ModalHeader>
-            <ModalBody>
-              <Tabs
-                tabs={TabOptions}
-                activeTab={activeTab}
-                onClick={(id) => setActiveTab(id)}
-                type="default"
-              />
-              {activeTab === 'visualizationData' && (
-                <div>
-                  <TabDescription>
-                    <BodyText
-                      variant="body2"
-                      color={transparentize(0.5, colors.black[100])}
-                    >
-                      {widgetType !== 'json'
-                        ? t('export_CSV.visualization_data_info')
-                        : t(
-                            'export_CSV.visualization_data_not_available_for_json'
-                          )}
-                    </BodyText>
-                  </TabDescription>
-                  {widgetType !== 'json' && (
-                    <Table
-                      data={visualizationData}
-                      columnLimit={3}
-                      rowLimit={3}
-                    />
-                  )}
-                </div>
-              )}
-              {activeTab === 'rawData' && (
-                <div>
-                  <TabDescription>
-                    <BodyText
-                      variant="body2"
-                      color={transparentize(0.5, colors.black[100])}
-                    >
-                      {t('export_CSV.raw_data_info')}
-                    </BodyText>
-                  </TabDescription>
-                  <Table data={rawData} columnLimit={6} rowLimit={3} />
-                </div>
-              )}
-            </ModalBody>
-            <ModalFooter>
-              <FooterButtonsWrapper>
-                <Button
-                  data-testid="save-query"
-                  variant="secondary"
-                  style="solid"
-                  onClick={() => {
-                    onExportToCSV();
-                  }}
-                  isDisabled={
-                    widgetType === 'json' && activeTab === 'visualizationData'
-                  }
-                >
-                  {t('export_CSV.export_csv')}
-                </Button>
-                <Anchor
-                  onClick={onClose}
-                  color={colors.blue[500]}
-                  hoverColor={colors.blue[300]}
-                >
-                  {t('export_CSV.cancel')}
-                </Anchor>
-              </FooterButtonsWrapper>
-            </ModalFooter>
-          </>
-        )}
+        {() => <ExportToCSV />}
       </Modal>
     </Portal>
   );
