+++ src/components/ExportToCSVModal/components/ExportToCSV/index.tsx
@@ -1,0 +1,3 @@
+import ExportToCSV from './ExportToCSV';
+
+export { ExportToCSV };
