+++ src/components/ExportToCSVModal/components/index.ts
@@ -1,0 +1,4 @@
+import { Table } from './Table';
+import { ExportToCSV } from './ExportToCSV';
+
+export { Table, ExportToCSV };
